Split automate_training and compare_models into two scripts
